tweaks, removed stuff and fixed bugs

+++ commands/Adminstuff/prunemembers.js
@@ -1,35 +1,25 @@
-exports.run = async (bot, msg, params = []) => {
+exports.run = async (bot, msg, params = []) => { // eslint-disable-line consistent-return
   let days;
   if (!params[0]) {
     const mes = await msg.channel.sendMessage('Wir kicken heute also inaktive Member?\nVon wie vielen Tagen der Inaktivität sprechen wir hier denn?');
-    try {
-      const collected = await mes.channel.awaitMessages(m => m.author.id === msg.author.id, { time: 30000, maxMatches: 1, errors: ['time'] })
-      mes.delete();
-      days = parseInt(collected.first().content.split(' ')[0]);
-      if (!days) {
-        return msg.channel.sendMessage('Bei dieser Eingabe handelt es sich um keine gültige Zahl!');
-      }
-    } catch (e) {
-      msg.delete();
-      return mes.delete();
-    }
+    const collected = (await mes.channel.awaitMessages(m => m.author.id === msg.author.id, { time: 30000, maxMatches: 1 })).first();
+    mes.delete();
+    if (!collected) return msg.delete();
+    days = parseInt(collected.content.split(' ')[0]);
+    if (!days) return msg.channel.sendMessage('Bei dieser Eingabe handelt es sich um keine gültige Zahl!');
   }
   const count = await msg.guild.pruneMembers(days, true);
   const mes = await msg.channel.sendMessage(`Das wären dann ${count} Member welche gekickt werden sollen.\nIst das korrekt? (__j__a/__y__es oder __n__ein/__n__o)`);
-  try {
-    const collected = await mes.channel.awaitMessages(m => m.author.id === msg.author.id, { time: 30000, maxMatches: 1, errors: ['time'] });
-    mes.delete();
-    if (['ja', 'yes', 'j', 'y'].includes(collected.first().content.split(' ')[0])) {
-      const kicked = await msg.guild.pruneMembers(days);
-      return msg.channel.sendMessage(`Es wurden erfolgreich ${kicked} Member gekickt!`);
-    } else if (['nein', 'no', 'n'].includes(collected.first().content.split(' ')[0])) {
-      return msg.channel.sendMessage('Breche den Vorgang dann, wie gewünscht, ab.');
-    } else {
-      return msg.channel.sendMessage('Konnte diese Eingabe nicht eindeutig zuordnen, breche den Vorgang daher vorsichtshabler ab.');
-    }
-  } catch (e) {
-    msg.delete();
-    return mes.delete();
+  const collected = (await mes.channel.awaitMessages(m => m.author.id === msg.author.id, { time: 30000, maxMatches: 1 })).first();
+  mes.delete();
+  if (!collected) return mes.delete();
+  if (['ja', 'yes', 'j', 'y'].includes(collected.content.split(' ')[0])) {
+    const kicked = await msg.guild.pruneMembers(days);
+    msg.channel.sendMessage(`Es wurden erfolgreich ${kicked} Member gekickt!`);
+  } else if (['nein', 'no', 'n'].includes(collected.content.split(' ')[0])) {
+    msg.channel.sendMessage('Breche den Vorgang dann, wie gewünscht, ab.');
+  } else {
+    msg.channel.sendMessage('Konnte diese Eingabe nicht eindeutig zuordnen, breche den Vorgang daher vorsichtshalber ab.');
   }
 };
